Fix iteration after multiple prepends
Previously, I assumed that you only need to reset the
current pointer in if the linked list's head was None. However,
if you prepend multiple times in a row, that leaves the current
pointer to be set to the value of the first prepended value. Now,
the current pointer is always reset to the head whenever we prepend
to the linked list.